% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BayesMCPMod}
\alias{BayesMCPMod}
\title{BayesMCPMod}
\usage{
BayesMCPMod(ancova, contMat1, crit_prob)
}
\arguments{
\item{ancova}{tbd}

\item{contMat1}{tbd}

\item{crit_prob}{tbd}
}
\value{
tbd
}
\description{
BayesMCPMod
}
