% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{BMCPMod}
\alias{BMCPMod}
\title{BMCPMod}
\usage{
BMCPMod(ancova1, cont_Mat1, crit_prob, n_simulations)
}
\arguments{
\item{ancova1}{tbd}

\item{cont_Mat1}{tbd}

\item{crit_prob}{tbd}

\item{n_simulations}{tbd}
}
\description{
BMCPMod
}
