% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{estimateModels}
\alias{estimateModels}
\title{estimateModels}
\usage{
estimateModels(models, post, doses)
}
\arguments{
\item{models}{tbd}

\item{post}{tbd}

\item{doses}{tbd}
}
\description{
estimateModels
}
