% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{getContr}
\alias{getContr}
\title{getContr}
\usage{
getContr(
  mods,
  dose_levels,
  dose_weights = NULL,
  prior_list = NULL,
  se_new_trial = NULL,
  sd_posterior = NULL
)
}
\arguments{
\item{mods}{tbd}

\item{dose_levels}{tbd}

\item{dose_weights}{tbd}

\item{prior_list}{tbd}

\item{se_new_trial}{tbd}

\item{sd_posterior}{tbd}
}
\description{
getContr
}
