% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{getContrMat}
\alias{getContrMat}
\title{getContrMat}
\usage{
getContrMat(mods, dose_levels, dose_weights, prior_list)
}
\arguments{
\item{mods}{tbd}

\item{dose_levels}{tbd}

\item{dose_weights}{tbd}

\item{prior_list}{tbd}
}
\description{
getContrMat
}
