% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{getCritProb}
\alias{getCritProb}
\title{getCritProb}
\usage{
getCritProb(mods, dose_levels, dose_weights, alpha_crit_val)
}
\arguments{
\item{mods}{tbd}

\item{dose_levels}{tbd}

\item{dose_weights}{tbd}

\item{alpha_crit_val}{tbd}
}
\description{
getCritProb
}
