% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPriorList}
\alias{getPriorList}
\title{getPriorList}
\usage{
getPriorList(hist_data, dose_levels, dose_names = NULL, robustify_weight = 0.5)
}
\arguments{
\item{hist_data}{tbd}

\item{dose_levels}{tbd}

\item{dose_names}{prior_list}

\item{robustify_weight}{tbd}
}
\description{
getPriorList
}
