% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCPMod}
\alias{performBayesianMCPMod}
\title{performBayesianMCPMod}
\usage{
performBayesianMCPMod(posterior_list, contr, crit_prob_adj, simple = FALSE)
}
\arguments{
\item{posterior_list}{a getPosterior object}

\item{contr}{a getContrMat object, contrast matrix to be used for the testing step.}

\item{crit_prob_adj}{a getCritProb object}

\item{simple}{boolean variable, defining whether simplified fit will be applied. Passed to the getModelFits function. Default FALSE.}
}
\value{
bmcpmod test result as well as modelling result.
}
\description{
performs bayesian MCP Test step and modelling.
}
