% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.modelFits}
\alias{plot.modelFits}
\title{plot.modelFits}
\usage{
\method{plot}{modelFits}(
  x,
  gAIC = TRUE,
  avg_fit = TRUE,
  cr_intv = TRUE,
  alpha_CrI = 0.05,
  cr_bands = FALSE,
  alpha_CrB = c(0.05, 0.5),
  n_bs_smpl = 1000,
  acc_color = "orange",
  ...
)
}
\arguments{
\item{x}{tbd An object of type modelFits}

\item{gAIC}{tbd}

\item{avg_fit}{tbd}

\item{cr_intv}{tbd}

\item{alpha_CrI}{tbd}

\item{cr_bands}{tbd}

\item{alpha_CrB}{tbd}

\item{n_bs_smpl}{tbd}

\item{acc_color}{tbd}

\item{...}{tbd}
}
\value{
tbd
}
\description{
plot.modelFits
}
