% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modelFits}
\alias{plot_modelFits}
\title{plot_modelFits}
\usage{
plot_modelFits(model_fits, CrI = FALSE, gAIC = TRUE, avg_fit = TRUE)
}
\arguments{
\item{model_fits}{tbd}

\item{CrI}{tbd}

\item{gAIC}{tbd}

\item{avg_fit}{tbd}
}
\value{
tbd
}
\description{
plot_modelFits
}
