% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSimulation.R
\name{simulateData}
\alias{simulateData}
\title{simulateData}
\usage{
simulateData(
  n_patients,
  dose_levels,
  pcov,
  SD,
  placebo_effect,
  max_effect,
  mod,
  n_simulations
)
}
\arguments{
\item{n_patients}{tbd}

\item{dose_levels}{tbd}

\item{pcov}{tbd}

\item{SD}{tbd}

\item{placebo_effect}{tbd}

\item{max_effect}{tbd}

\item{mod}{tbd}

\item{n_simulations}{tbd}
}
\description{
simulateData
}
