% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{assessDesign}
\alias{assessDesign}
\title{assessDesign}
\usage{
assessDesign(
  n_patients,
  mods,
  prior_list,
  n_sim = 1000,
  alpha_crit_val = 0.05,
  simple = TRUE
)
}
\arguments{
\item{n_patients}{tbd}

\item{mods}{tbd}

\item{prior_list}{tbd}

\item{n_sim}{tbd}

\item{alpha_crit_val}{tbd}

\item{simple}{tbd}
}
\description{
assessDesign
}
