% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{getBootstrapQuantiles}
\alias{getBootstrapQuantiles}
\title{getBootstrapQuantiles}
\usage{
getBootstrapQuantiles(
  model_fits,
  quantiles,
  n_samples = 1000,
  doses = NULL,
  avg_fit = TRUE
)
}
\arguments{
\item{model_fits}{tbd}

\item{n_samples}{tbd}

\item{doses}{tbd}

\item{avg_fit}{tbd}

\item{alpha}{tbd}
}
\value{
tbd
}
\description{
getBootstrapQuantiles
}
