% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{getBootstrapQuantiles}
\alias{getBootstrapQuantiles}
\title{getBootstrapQuantiles}
\usage{
getBootstrapQuantiles(
  model_fits,
  n_samples = 1000,
  alpha = c(0.05, 0.5),
  avg_fit = TRUE,
  dose_seq = NULL
)
}
\arguments{
\item{model_fits}{tbd}

\item{n_samples}{tbd}

\item{alpha}{tbd}

\item{avg_fit}{tbd}

\item{dose_seq}{tbd}
}
\value{
tbd
}
\description{
getBootstrapQuantiles
}
