% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{getBootstrapQuantiles}
\alias{getBootstrapQuantiles}
\title{getBootstrapQuantiles}
\usage{
getBootstrapQuantiles(
  model_fits,
  quantiles,
  n_samples = 1000,
  doses = NULL,
  avg_fit = TRUE
)
}
\arguments{
\item{model_fits}{an object of class modelFits, i.e. information about fitted models & corresponding model coefficients as well as the posterior distribution that was the basis for the model fitting}

\item{quantiles}{a vector of quantiles that should be evaluated}

\item{n_samples}{number of samples that should be drawn as basis for the}

\item{doses}{a vector of doses for which a prediction should be performed}

\item{avg_fit}{boolean variable, defining whether an average fit (based on generalized AIC weights) should be performed in addition to the individual models. Default TRUE.}
}
\value{
A data frame with entries doses, models, and quantiles
}
\description{
A function for the calculation of credible intervals to assess the uncertainty for the model fit.
Hereby the credible intervals are calculated as follows.
Samples from the posterior distribution are drawn (via the RBesT function rmix) and for every sample the simplified fitting step (see getModelFits function) and a prediction is performed.
These fits are then used to identify the specified quantiles.
This approach can be considered as the bayesian equivalent of the frequentist bootstrap approach described in \link{O'Quigley et al., 2017}.
Instead of drawing n bootstrap samples from the sampling distribution of the trial dose-response estimates, here the samples are directly taken from the posterior distribution.
}
\examples{
# example code
posterior_list =   list(Ctrl=RBesT::mixnorm(comp1 = c(w = 1, m = 0, s = 1), sigma = 2),
                   DG_1=RBesT::mixnorm(comp1 = c(w = 1, m = 3, s = 1.2), sigma = 2),
                   DG_2=RBesT::mixnorm(comp1 = c(w = 1, m = 4, s = 1.5), sigma = 2) ,  
                   DG_3=RBesT::mixnorm(comp1 = c(w = 1, m = 6, s = 1.2), sigma = 2) ,
                   DG_4=RBesT::mixnorm(comp1 = c(w = 1, m = 6.5, s = 1.1) ,sigma = 2))
models=c("emax","exponential","sigEmax","linear")
dose_levels=c(0,1,2,4,8)
fit<-getModelFits(models=models, posterior=posterior_list,dose_levels=dose_levels)
fit_simple<-getModelFits(models=models, posterior=posterior_list,dose_levels=dose_levels,simple=TRUE)
getBootstrapQuantiles(fit, quantiles = c(0.025,0.5, 0.975), doses = c(0, 1,2,3,4,6,8))
getBootstrapQuantiles(fit_simple, n_samples=2000, quantiles = c(0.025,0.5, 0.975), doses = c(0, 1,2,3,4,6,8))
}
\references{
O'Quigley, J., Iasonos, A., & Bornkamp, B. (Eds.). (2017). Handbook of Methods for Designing, Monitoring, and Analyzing Dose-Finding Trials (1st ed.). Chapman and Hall/CRC. https://doi.org/10.1201/9781315151984
}
