% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{getContrMat}
\alias{getContrMat}
\title{getContr}
\usage{
getContrMat(
  mods,
  dose_levels,
  dose_weights = NULL,
  prior_list = NULL,
  se_new_trial = NULL,
  sd_posterior = NULL
)
}
\arguments{
\item{mods}{An object of class "Mods" as specified in the Dosefinding package.}

\item{dose_levels}{vector containing the different doseage levels.}

\item{dose_weights}{Vector specifying weights for the different doses}

\item{prior_list}{a prior_list object}
}
\value{
contr_mat Object of class ‘⁠optContr⁠’. A list containing entries contMat and muMat, and CorrMat. Specified in the Dosefinding package.
}
\description{
This function calculates contrast vectors that are optimal for detecting certain alternatives. More information and link to publication will be added.
}
