% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getESS}
\alias{getESS}
\title{getESS}
\usage{
getESS(post_list)
}
\arguments{
\item{post_list}{a posterior list object, for which the effective sample size (per dose group) should be calculated}
}
\value{
a vector of the effective sample sizes (per dose group)
}
\description{
This function calculates the effective sample size for every dose group via the RBesT function ess.
}
