% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{getModelFits}
\alias{getModelFits}
\title{getModelFits}
\usage{
getModelFits(models, dose_levels, posterior, simple = FALSE)
}
\arguments{
\item{models}{tbd}

\item{dose_levels}{tbd}

\item{posterior}{tbd}

\item{simple}{tbd}
}
\description{
getModelFits
}
