% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD
getPosterior(data = NULL, prior_list, mu_hat = NULL, sd_hat = NULL)
}
\arguments{
\item{data}{dataframe containing the information of dose and response.
Also a simulateData object can be provided.}

=======
getPosterior(
  prior_list,
  data = NULL,
  mu_hat = NULL,
  se_hat = NULL,
  calc_ess = FALSE
)
}
\arguments{
>>>>>>> 25f8f28541ce5be0c55fecee4cdd47a6c8603237
\item{prior_list}{prior_list object}

\item{data}{dataframe containing the information of dose and response. Default NULL
Also a simulateData object can be provided.}

\item{mu_hat}{vector of estimated mean values (per dose group).}

<<<<<<< HEAD
<<<<<<< HEAD
\item{sd_hat}{vector of estimated standard deviations.}
=======
getPosterior(prior_list, data = NULL, mu_hat = NULL, se_hat = NULL)
}
\arguments{
\item{prior_list}{prior_list}

\item{data}{tbd}

\item{mu_hat}{tbd}

\item{se_hat}{tbd}
>>>>>>> ff9a940b072d43dcf445d6c6944d4b8bcbc07a8b
=======
\item{se_hat}{vector of estimated standard deviations.}

\item{calc_ess}{tbd. Default NULL}
>>>>>>> 25f8f28541ce5be0c55fecee4cdd47a6c8603237
=======
\item{se_hat}{vector of estimated standard deviations (per dose group).}

\item{calc_ess}{boolean variable, indicating whether effective sample size should be calculated. Default FALSE}
}
\value{
posterior_list, a posterior list object is returned with information about (mixture) posterior distribution per dose group
>>>>>>> cb4cf39ef75ca89ecdb587b5ba0b0c863745f760
}
\description{
Either the patient level data or both mu_hat as well as sd_hat must to be provided. If patient level data is provided mu_hat and se_hat are calculated within the function using a linear model.
This function calculates the posterior for every dose group independently via the RBesT function postmix.
}
