% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
getPosterior(prior_list, data = NULL, mu_hat = NULL, se_hat = NULL)
}
\arguments{
\item{prior_list}{prior_list}

\item{data}{tbd}

\item{mu_hat}{tbd}

\item{se_hat}{tbd}
}
\description{
getPosterior
}
