% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
getPosterior(prior_list, data = NULL, mu_hat = NULL, se_hat = NULL)
}
\arguments{
\item{prior_list}{prior_list object}

\item{data}{dataframe containing the information of dose and response.
Also a simulateData object can be provided.}

\item{mu_hat}{vector of estimated mean values}

\item{sd_hat}{vector of estimated standard deviations.}
}
\description{
Either the patient level data or both the mu_hat as well as the sd_hat must to be provided.
}
