% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
getPosterior(data, prior_list, mu_hat = NULL, sd_hat = NULL)
}
\arguments{
\item{data}{tbd}

\item{prior_list}{prior_list}

\item{mu_hat}{tbd}

\item{sd_hat}{tbd}
}
\description{
getPosterior
}
