% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPriorList}
\alias{getPriorList}
\title{getPriorList}
\usage{
getPriorList(hist_data, dose_levels, dose_names = NULL, robustify_weight)
}
\arguments{
\item{hist_data}{historical trial summary level data,
needs to be provided as a dataframe. Including information of the
estimates and variability.}

\item{dose_levels}{vector of the different doseage levels}

\item{dose_names}{character vector of dose levels,
default NULL and will be automatically created
based on the dose levels parameter.}

\item{robustify_weight}{needs to be provided as a numeric
value for the weight of the robustification component}
}
\description{
getPriorList
}
