% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCP}
\alias{performBayesianMCP}
\title{performBayesianMCP}
\usage{
performBayesianMCP(posteriors_list, contr_mat, crit_prob_adj)
}
\arguments{
\item{posteriors_list}{tbd}

\item{contr_mat}{tbd}

\item{crit_prob_adj}{tbd}
}
\description{
performBayesianMCP
}
