% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCP}
\alias{performBayesianMCP}
\title{performBayesianMCP}
\usage{
performBayesianMCP(posterior_list, contr, crit_prob_adj)
}
\arguments{
\item{posterior_list}{a getPosterior object}

\item{contr}{a getContrMat object, contrast matrix to be used for the testing step.}

\item{crit_prob_adj}{a getCritProb object, specifying the critical value to be used for the testing (on the probability scale)}
}
\value{
b_mcp test result
}
\description{
performs bayesian MCP Test step.
}
