% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCP}
\alias{performBayesianMCP}
\title{performBayesianMCP}
\usage{
performBayesianMCP(posteriors_list, contr_mat, crit_prob)
}
\arguments{
\item{posteriors_list}{a getPosterior object}

\item{contr_mat}{a getContrMat object, contrast matrix to be used for the testing step.}

\item{crit_prob}{a getCritProb object, specifying the critical value to be used for the testing (on the probability scale)}
}
\value{
b_mcp test result
}
\description{
performs bayesian MCP Test step.
}
