% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCPMod}
\alias{performBayesianMCPMod}
\title{performBayesianMCPMod}
\usage{
performBayesianMCPMod(
  posteriors_list,
  contr_mat,
  crit_prob_adj,
  simple = FALSE
)
}
\arguments{
\item{posteriors_list}{tbd}

\item{contr_mat}{tbd}

\item{crit_prob_adj}{tbd}

\item{simple}{tbd}
}
\description{
performBayesianMCPMod
}
