% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMCPMod.R
\name{performBayesianMCPMod}
\alias{performBayesianMCPMod}
\title{performBayesianMCPMod}
\usage{
performBayesianMCPMod(
  posteriors_list,
  contr_mat,
  crit_prob_adj,
  simple = FALSE
)
}
\arguments{
\item{posteriors_list}{a getPosterior object}

\item{contr_mat}{a getContrMat object, contrast matrix to be used for the testing step.}

\item{simple}{boolean variable, defining whether simplified fit will be applied. Passed to the getModelFits function. Default FALSE.}

\item{crit_prob}{a getCritProb object}
}
\value{
bmcpmod test result as well as modelling result.
}
\description{
performs bayesian MCP Test step and modelling.
}
