% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.modelFits}
\alias{plot.modelFits}
\title{plot.modelFits}
\usage{
\method{plot}{modelFits}(
  x,
  gAIC = TRUE,
  avg_fit = TRUE,
  cr_intv = TRUE,
  alpha_CrI = 0.05,
  cr_bands = FALSE,
  alpha_CrB = c(0.05, 0.5),
  n_bs_smpl = 1000,
  acc_color = "orange",
  ...
)
}
\arguments{
\item{x}{an object of type getModelFits}

\item{gAIC}{logical value indicating whether gAIC values are shown in the plot. Default TRUE}

\item{avg_fit}{logical value indicating whether average fit is presented in the plot. Default TRUE}

\item{cr_intv}{logical value indicating whether credible intervals are included in the plot. Default TRUE}

\item{alpha_CrI}{numerical value of the width of the credible intervals. Default is set to 0.05 (i.e 95\% CI are shown).}

\item{cr_bands}{logical value indicating whether bootstrapped based credible bands are shown in the plot. Default FALSE}

\item{alpha_CrB}{numerical vector of the width of the credible bands. Default is set to 0.05 and 0.5 (i.e 95\% CB and median are shown).}

\item{n_bs_smpl}{number of bootstrap samples being used. Default set to 1000.}

\item{acc_color}{color of the credible bands. Default set to "orange"}

\item{...}{optional parameter to be passed.}
}
\value{
plts returns a ggplot2 object
}
\description{
plot function based on the ggplot2 package. Providing visualizations for each model and a average Fit.
More details to be added, as well as references.
}
