% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modelFits}
\alias{plot_modelFits}
\title{plot_modelFits}
\usage{
plot_modelFits(
  model_fits,
  gAIC = TRUE,
  avg_fit = TRUE,
  cr_intv = FALSE,
  alpha_CrI = 0.05,
  cr_bands = FALSE,
  alpha_CrB = c(0.05, 0.5),
  n_bs_smpl = 1000
)
}
\arguments{
\item{model_fits}{tbd}

\item{gAIC}{tbd}

\item{avg_fit}{tbd}

\item{cr_intv}{tbd}

\item{alpha_CrI}{tbd}

\item{cr_bands}{tbd}

\item{alpha_CrB}{tbd}

\item{n_bs_smpl}{tbd}
}
\value{
tbd
}
\description{
plot_modelFits
}
