% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateData}
\alias{simulateData}
\title{simulateData}
\usage{
simulateData(n_patients, dose_levels, sd, mods, n_sim)
}
\arguments{
\item{n_patients}{tbd}

\item{dose_levels}{tbd}

\item{sd}{tbd}

\item{mods}{tbd}

\item{n_sim}{tbd}
}
\description{
simulateData
}
