% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateData}
\alias{simulateData}
\title{simulateData}
\usage{
simulateData(
  n_patients,
  dose_levels,
  sd,
  mods,
  n_sim = 1000,
  true_model = NULL,
  dr_means = NULL
)
}
\arguments{
\item{n_patients}{vector containing number of patients as a numerical
value per dose-group.}

\item{dose_levels}{vector containing the different doseage levels.}

\item{sd}{standard deviation on patient level.}

\item{mods}{An object of class "Mods" as specified in the Dosefinding package.}

\item{n_sim}{number of simulations to be performed,
Default is 1000}

\item{true_model}{Default value is NULL.
Assumed true underlying model. Provided via a String. e.g. "emax".
In case of NULL, all dose-response models, included in the mods input parameter will be used.}
}
\value{
sim_data one list object, containing patient level simulated data for all assumed true models.
Also providing information about simulation iteration, patient number as well as dosage levels.
}
\description{
simulateData
}
